Add .editorconfig file, standardize formatting
I see quite a of bit of inconsistency in the formatting.  Let's add a
.editorconfig file, so the code we write uses the right whitespacing
regardless of the editor we use.  I just copied the one I wrote earlier
for Theia to get started, but it doesn't have to be identical.

I used Theia to auto-format all ts and json files and then tslint --fix
to fixup any warnings that it would have inserted.

For example, I the Typescript formatter handles quite badly (IMO) wrapping long
parameter lists, and produces code that result in a warning according to
our current tslint config.  The Typescript formatter will do this:

```
export class Bob {
    constructor(a: number, b: number, c: number, d: number,
        e: number, f: number) {
        console.log('Hello');
    }
}
```

where the second line of parameters is indented with 4 spaces.  I think
is not very readable, because it's easily confused with the first line
of code.  I prefer our current tslint setting, which requires it to be
aligned like this:

```
export class Bob {
    constructor(a: number, b: number, c: number, d: number,
i               e: number, f: number) {
        console.log('Hello');
    }
}
```

This means that formatting entire documents will always change the code
to look like the first snippet above.  It's a minor annoyance, and
hopefully there will be a setting for that one day in the Typescript
formatter.  There is an issue about that here:

https://github.com/Microsoft/TypeScript/issues/11629
Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>